feat: Email confirmation to login

+++ apps/accounts/src/app/pages/sign-in/sign-in.ts
@@ -1,6 +1,6 @@
 import { CommonModule } from '@angular/common';
 import { HttpClient } from '@angular/common/http';
-import { Component, inject, OnInit } from '@angular/core';
+import { Component, inject, OnInit, signal } from '@angular/core';
 import {
   FormControl,
   FormGroup,
@@ -73,6 +73,9 @@ export class SignIn implements OnInit {
     password: new FormControl('Teste@1234', [Validators.required]),
     recaptcha: new FormControl('', []),
   });
+  isEmailNotConfirmed = signal<boolean>(false);
+  disableSubmit = signal<boolean>(false);
+  isLoading = signal<boolean>(false);
 
   router = inject(Router);
   activatedRoute = inject(ActivatedRoute);
@@ -129,13 +132,14 @@ export class SignIn implements OnInit {
 
   async signInWithExternalOidc() {
     try {
-      const { idToken, isAuthenticated, accessToken, ...all } = await firstValueFrom(
-        this.oidcSecurityService.authorizeWithPopUp(
-          undefined,
-          undefined,
-          EXTERNAL_OIDC_AUTH_CONFIG_ID,
-        ),
-      );
+      const { idToken, isAuthenticated, accessToken, ...all } =
+        await firstValueFrom(
+          this.oidcSecurityService.authorizeWithPopUp(
+            undefined,
+            undefined,
+            EXTERNAL_OIDC_AUTH_CONFIG_ID,
+          ),
+        );
       console.log('all', all);
       console.log('accessToken', accessToken);
 
@@ -157,6 +161,15 @@ export class SignIn implements OnInit {
 
           this.redirectOnSuccess(response);
         } catch (err: any) {
+          if (err.status === 400 && err.error?.isEmailNotConfirmed) {
+            this.isEmailNotConfirmed.set(true);
+            this.disableSubmit.set(true);
+
+            setTimeout(() => {
+              this.disableSubmit.set(false);
+            }, 5000);
+            return;
+          }
           if (err.status === 401 && err.error?.message === 'USER_NOT_FOUND') {
             localStorage.setItem(
               'govBrTokens',
@@ -167,11 +180,16 @@ export class SignIn implements OnInit {
                 timestamp: Date.now(),
               }),
             );
-            this.toaster.show(this.translate.instant('pages.signIn.errors.userNotFound'), { type: 'info' });
+            this.toaster.show(
+              this.translate.instant('pages.signIn.errors.userNotFound'),
+              { type: 'info' },
+            );
             this.router.navigate(['/sign-up']);
             return;
           }
-          this.toaster.error(this.translate.instant('pages.signIn.errors.submit'));
+          this.toaster.error(
+            this.translate.instant('pages.signIn.errors.submit'),
+          );
           throw err;
         }
       }
@@ -186,6 +204,7 @@ export class SignIn implements OnInit {
   }
 
   async onSubmit() {
+    this.isLoading.set(true);
     this.formGroup.controls.email.setValue(
       this.formGroup.controls.email.value?.replace(/\s/g, '') ?? '',
       { emitEvent: false },
@@ -198,10 +217,24 @@ export class SignIn implements OnInit {
     console.log('formGroup.value', this.formGroup.value);
 
     this.signInService.authenticate(this.formGroup.value).subscribe({
-      error: ({ error }) => {
-        console.error(error);
+      error: (err) => {
+        this.isLoading.set(false);
+        console.error(err);
+
+        if (err.status === 400 && err.error?.isEmailNotConfirmed) {
+          this.isEmailNotConfirmed.set(true);
+          this.disableSubmit.set(true);
+
+          setTimeout(() => {
+            this.disableSubmit.set(false);
+          }, 5000);
+          return;
+        }
       },
-      next: (res: any) => this.redirectOnSuccess(res),
+      next: (res: any) => {
+        this.isLoading.set(false);
+        this.redirectOnSuccess(res);
+      },
     });
   }
